requirements.txt
click==8.1.7
Flask==2.2.5
flask-sqlalchemy==3.0.5
greenlet==3.0.2
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
typing-extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0
